Code style refactor. Add description and RPQ rename
